Changed the constant register from -1 to 0
I originally conceived the N register in an instruction set
that lacked immediate values. In the instruction set, -1 was
immensly useful for fabricating instructions such as branch always,
not, decrement, increment, nop, etc. However, c16 makes heavy use
of immediate values. This removes most of the benefits of constant
registers in general, but c16 cannot remove its constant register
without losing the ability to unconditionally branch.

I was content to just leave the constant register as -1 simply because
every use of 0 easily mapped to a use of -1, until I found these two uses:

sub z,ra,rd     => neg ra,rd (negate)
sltu z,ra,rd    => sne ra,rd (set if not zero)

As far as I'm aware, these cannot be mapped to instructions using -1.
It's not much of an addition to the instruction set, but there is
no reason that I'm aware of to keep using -1.