Added software interrupts and extension opcodes
The software interrupt instruction `int` provides a mechanism to
interact with an operating system without hacky mechanics. It does
not necessarily require special decoding as the interrupt normally
disrupts the flow of the processor.

The extension opcode is a normally undefined opcode that allows
application specific instructions to be defined in the lower bits of
the instruction. This opcode should only be used for processor specific
operations, and should not be used by portable applications. This
adds flexibility to the instruction set and provides a method for
prototyping new instructions.