Created 32-bit version of c16
This is not intended for use in class but is just an extension of
c16 to 32-bit words with minimum changes to the instruction set itself.
If anyone wants to use all of the available memory without using another
instruction set, this should be sufficient.

The biggest change is the removal of the logical shift right. 32-bit
words now requires 5-bits to specify the shift and the instruction
does not have the space for the arith/logic bit. If a logical shift
is needed, a mask can be used with an arithmetic shift.

Additionally, the PC and other registers are now byte-addressable.
This doesn't change the fact that unaligned word access can be
unallowed, but there is no reason to not include these bits with
a 32-bit address space. All address offsets now use an implicit
shift to be able to changes addresses by 16-bit intervals. This
makes the instruction set's branches/jumps identical to the c16.